_get_hours_per_day_list now robust, handles work_today
spec clarification: when not working today and there are
tasks due tomorrow, we silently ignore those tasks. we
may later want to raise an exception in this case, turning
making the user work today until the issue is cleared up.
This was found using Hypothesis testing. A much more
sophisticated model for _get_hours_per_day_list is in place
now in this commit.